Remove BLMainServer
