+++ src/com/blazeloader/api/main/BLMainClient.java
@@ -8,7 +8,7 @@ import net.minecraft.client.Minecraft;
  * Client BLMain.
  */
 public class BLMainClient extends BLMain {
-    public BLMainClient(LoaderEnvironment environment, LoaderProperties properties) {
+    BLMainClient(LoaderEnvironment environment, LoaderProperties properties) {
         super(environment, properties);
     }
 
@@ -19,23 +19,22 @@ public class BLMainClient extends BLMain {
 
     @Override
     public void shutdown(String message, int code) {
-        LOGGER_FULL.logFatal("Unexpected shutdown requested!");
-        LOGGER_FULL.logFatal("Message: " + message);
-        Minecraft minecraft = Minecraft.getMinecraft();
-        if (minecraft != null) {
-            LOGGER_FULL.logFatal("Calling client shutdown.");
-            minecraft.shutdown();
-        } else {
-            LOGGER_FULL.logFatal("Client is not running, closing immediately with code " + code + "!");
+        try {
+            LOGGER_FULL.logFatal("Unexpected shutdown requested!");
+            LOGGER_FULL.logFatal("Message: " + message);
+            Minecraft minecraft = Minecraft.getMinecraft();
+            if (minecraft != null) {
+                LOGGER_FULL.logFatal("Calling client shutdown.");
+                minecraft.shutdown();
+            } else {
+                LOGGER_FULL.logFatal("Client is not running, closing immediately with code " + code + "!");
+                System.exit(code);
+            }
+        } catch (Throwable t) {
+            t.printStackTrace();
             System.exit(code);
         }
     }
-
-    @Override
-    public boolean supportsServer() {
-        return false;
-    }
-
     @Override
     public boolean supportsClient() {
         return true;
@@ -44,10 +43,5 @@ public class BLMainClient extends BLMain {
     @Override
     public BLMainClient getClient() {
         return this;
-    }
-
-    @Override
-    public BLMainServer getServer() {
-        throw new UnsupportedOperationException("This BLMain does not support BLMainServe!");
     }
 }
