+++ src/com/blazeloader/api/version/Versions.java
@@ -31,16 +31,7 @@ public class Versions {
      * @return Return true is running on client, false for server.
      */
     public static boolean isClient() {
-        return !isServer();
-    }
-
-    /**
-     * Returns true if running on the server, false for client.  Only works AFTER BlazeLoader is initialized.
-     *
-     * @return Return true is running on server, false for client.
-     */
-    public static boolean isServer() {
-        return BLMain.instance().supportsServer();
+        return BLMain.instance().supportsClient();
     }
 
     /**
